pyproject.toml
[project]
name = "militol-audio-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "certifi>=2025.10.5",
    "fastapi>=0.115.0",
    "demucs>=4.0.1",
    "kagglehub>=0.3.13",
    "librosa>=0.11.0",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "openai>=2.6.1",
    "openai-whisper>=20250625",
    "pandas>=2.3.3",
    "pyannote-audio>=4.0.1",
    "pydantic-settings>=2.11.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.2",
    "soundfile>=0.13.1",
    "tensorboard>=2.20.0",
    "torch>=2.9.0",
    "torchaudio>=2.9.0",
    "torchcodec>=0.8.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
    "uvicorn>=0.32.0",
]
